requirements.txt
paho-mqtt==2.1.0
pandas==2.2.2
python-dotenv==1.0.1
streamlit==1.51.0
folium==0.20.0
streamlit-folium==0.25.3
